Several bug fixes - major commit.
1. The app would crash because the build target was not set to Google Apis, and
because the uses-library in AndroidManifest.xml was not within the
<Application> tag. Fixed both of these. However, had to set minSDK to 3. Not
sure if this is required - seems unreasonable.

2. Several NullPointerExceptions were being caused by the fact that
savedInstanceState was not being checked for null. Fixed this.

3. When control returned to Home screen after creating a new alarm, the alarm
list would not be refreshed. This was due to two reasons. First, when control
returns, onCreate is not called, only onResume is. Moved the setupUI() call to
onResume to fix this. Second, Alarms was designed assuming that there would
only be one Alarms objects (Singleton). However designed changed since there is
a need to pass in a Context for Alarms. The result was if an alarm was updated
in one Alarms object, the other would not know about it. The fix was to make
Alarms read the Alarms file in on every function call to it.